+++ Microsoft.Toolkit.Services/Services/MicrosoftGraph/MicrosoftGraphAuthenticationHelper.cs
@@ -10,10 +10,8 @@ using System.Threading.Tasks;
 using Microsoft.Identity.Client;
 
 #if WINRT || WINDOWS_UWP
-using Microsoft.IdentityModel.Clients.ActiveDirectory;
 using Windows.Storage;
 #endif
-using MSAL = Microsoft.Identity.Client;
 
 namespace Microsoft.Toolkit.Services.MicrosoftGraph
 {
@@ -26,7 +24,6 @@ namespace Microsoft.Toolkit.Services.MicrosoftGraph
         /// Base Url for service.
         /// </summary>
         private const string Authority = "https://login.microsoftonline.com/{0}/";
-
 
         /// <summary>
         /// Default Logout Url for V2
@@ -71,18 +68,15 @@ namespace Microsoft.Toolkit.Services.MicrosoftGraph
             _authHelperImplementation = new AuthenticationHelperV2(clientId, string.Format(Authority, tenantId), string.Format(LogoutUrlV2Model, tenantId));
         }
 
-
         /// <summary>
-        /// ClientId for the application
+        /// Gets ClientId for the application
         /// </summary>
         public string ClientId { get; }
-
 
         public Task<string> AquireTokenAsync(IEnumerable<string> scopes, UIParent uiParent = null, string redirectUri = null, string loginHint = null)
         {
             return _authHelperImplementation.AquireTokenAsync(scopes, uiParent, redirectUri, loginHint);
         }
-
 
         /// <summary>
         /// Logout the user
@@ -95,144 +89,6 @@ namespace Microsoft.Toolkit.Services.MicrosoftGraph
 #endif
 
             return await _authHelperImplementation.LogoutAsync().ConfigureAwait(false);
-        }
-
-        public Task<string> AquireTokenAsync(string resourceId, UIParent uiParent = null, string redirectUri = null, string loginHint = null)
-        {
-            return AquireTokenAsync(new[] { resourceId }, uiParent, redirectUri, loginHint);
-        }
-    }
-
-    public interface IAuthenticationHelper
-    {
-        Task<string> AquireTokenAsync(string resourceId, UIParent uiParent = null, string redirectUri = null, string loginHint = null);
-
-        Task<string> AquireTokenAsync(IEnumerable<string> scopes, UIParent uiParent = null, string redirectUri = null, string loginHint = null);
-
-        Task<bool> LogoutAsync();
-    }
-
-#if WINRT || WINDOWS_UWP
-    internal class AuthenticationHelperV1 : IAuthenticationHelper
-    {
-        /// <summary>
-        /// Default Redirect Uri
-        /// </summary>
-        private const string DefaultRedirectUri = "urn:ietf:wg:oauth:2.0:oob";
-
-        private readonly AuthenticationContext _authContext;
-        private readonly string clientId;
-        private readonly string logoutUrl;
-
-        public AuthenticationHelperV1(string clientId, string authority, string logoutUrl)
-        {
-            _authContext = new AuthenticationContext(authority);
-            this.clientId = clientId;
-            this.logoutUrl = logoutUrl;
-        }
-
-
-
-        public async Task<string> AquireTokenAsync(IEnumerable<string> scopes, UIParent uiParent = null, string redirectUri = null, string loginHint = null)
-        {
-            try
-            {
-                IdentityModel.Clients.ActiveDirectory.AuthenticationResult userAuthnResult = await _authContext.AcquireTokenSilentAsync(scopes.First(), clientId);
-                return userAuthnResult.AccessToken;
-            }
-            catch (AdalSilentTokenAcquisitionException e)
-            {
-                try
-                {
-                    IdentityModel.Clients.ActiveDirectory.AuthenticationResult userAuthnResult = await _authContext.AcquireTokenAsync(scopes.First(), clientId, new Uri(redirectUri ?? DefaultRedirectUri), new IdentityModel.Clients.ActiveDirectory.PlatformParameters(PromptBehavior.SelectAccount, false));
-                    return userAuthnResult.AccessToken;
-                }
-                catch (AdalException exception)
-                {
-                    throw;
-                }
-            }   
-        }
-
-        public Task<string> AquireTokenAsync(string resourceId, UIParent uiParent = null, string redirectUri = null, string loginHint = null)
-        {
-            return AquireTokenAsync(new[] { resourceId }, uiParent, redirectUri, loginHint);
-        }
-
-        public async Task<bool> LogoutAsync()
-        {
-            using (var client = new HttpClient())
-            {
-                var request = new HttpRequestMessage(HttpMethod.Get, logoutUrl);
-                var response = await client.SendAsync(request);
-                return response.IsSuccessStatusCode;
-            }
-        }
-    }
-#endif
-
-    internal class AuthenticationHelperV2 : IAuthenticationHelper
-    {
-        private readonly PublicClientApplication _identityClient;
-        private readonly string logoutUrl;
-
-        public AuthenticationHelperV2(string clientId, string authority, string logoutUrl)
-        {
-            _identityClient = new PublicClientApplication(clientId, authority);
-            this.logoutUrl = logoutUrl;
-        }
-
-        public async Task<string> AquireTokenAsync(IEnumerable<string> scopes, UIParent uiParent = null, string redirectUri = null, string loginHint = null)
-        {
-            if (!string.IsNullOrEmpty(redirectUri))
-            {
-                _identityClient.RedirectUri = redirectUri;
-            }
-
-            string upnLoginHint = null;
-            if (!string.IsNullOrEmpty(loginHint))
-            {
-                upnLoginHint = loginHint;
-            }
-
-            MSAL.AuthenticationResult authenticationResult = null;
-
-            try
-            {
-                authenticationResult = await _identityClient.AcquireTokenSilentAsync(scopes, _identityClient.Users.FirstOrDefault());
-            }
-            catch (MsalUiRequiredException)
-            {
-                try
-                {
-                    authenticationResult = await _identityClient.AcquireTokenAsync(scopes, upnLoginHint, UIBehavior.SelectAccount, null, uiParent);
-                }
-                catch (MsalException)
-                {
-                    throw;
-                }
-            }
-
-            return authenticationResult?.AccessToken;
-        }
-
-        public async Task<bool> LogoutAsync()
-        {
-            try
-            {
-                _identityClient.Remove(_identityClient.Users.FirstOrDefault());
-
-                using (var client = new HttpClient())
-                {
-                    var request = new HttpRequestMessage(HttpMethod.Get, logoutUrl);
-                    var response = await client.SendAsync(request);
-                    return response.IsSuccessStatusCode;
-                }
-            }
-            catch (MsalException)
-            {
-                return false;
-            }
         }
 
         public Task<string> AquireTokenAsync(string resourceId, UIParent uiParent = null, string redirectUri = null, string loginHint = null)
